Convert EmailController to EmailService
