+++ identity/app/views/profile/privacyForm.scala.html
@@ -1,14 +1,20 @@
-@(idUrlBuilder: services.IdentityUrlBuilder,
-  idRequest: services.IdentityRequest,
-  user: com.gu.identity.model.User,
-  privacyForm: Form[_root_.form.PrivacyFormData])(implicit request: RequestHeader, messages: play.api.i18n.Messages)
-
 @import views.html.fragments.form.checkboxField
 @import views.html.fragments.registrationFooter
 @import _root_.form.IdFormHelpers.{Checkbox, nonInputFields}
 @import com.gu.identity.model.ConsentText
 @import com.gu.identity.model.ConsentText._
 @import conf.switches.Switches.IdentityGdprMarketingConsentSwitch
+@import model.{ApplicationContext, EmailNewsletters, IdentityPage}
+@import services.EmailPrefsData
+
+@(idUrlBuilder: services.IdentityUrlBuilder,
+  idRequest: services.IdentityRequest,
+  user: com.gu.identity.model.User,
+  privacyForm: Form[_root_.form.PrivacyFormData],
+  emailPrefsForm: Form[EmailPrefsData],
+  emailSubscriptions: List[String],
+  availableLists: EmailNewsletters)(implicit request: RequestHeader, messages: play.api.i18n.Messages)
+
 
 @v2ConsentCheckbox(consentField: Field) = {
 
@@ -42,18 +48,19 @@
     @checkboxField(Checkbox(privacyForm("receive3rdPartyMarketing"), '_label -> currentConsents(ThirdParty.name)))(nonInputFields, messages)
 }
 
-<form class="form js-public-profile-form" novalidate action="@idUrlBuilder.buildUrl("/privacy/edit", idRequest)" role="main" method="post">
-    @views.html.helper.CSRF.formField
+<div>
+    <form class="form js-public-profile-form" novalidate action="@idUrlBuilder.buildUrl("/privacy/edit", idRequest)" role="main" method="post">
+        @views.html.helper.CSRF.formField
 
-    @* This displays both global and key-bound errors *@
-    @if(privacyForm.hasErrors) {
-        <div class="form__error">
-          Error processing the form. Your changes have not been saved:
-          <p>@privacyForm.errors.map(formError => (formError.key, formError.message)).mkString(", ")</p>
-        </div>
-    }
+        @* This displays both global and key-bound errors *@
+        @if(privacyForm.hasErrors) {
+            <div class="form__error">
+                Error processing the form. Your changes have not been saved:
+                <p>@privacyForm.errors.map(formError => (formError.key, formError.message)).mkString(", ")</p>
+            </div>
+        }
 
-    @* Newsletters *@
+        @* Newsletters *@
     <fieldset class="fieldset">
         <div class="fieldset__heading">
             <h2 class="form__heading">Newsletters</h2>
@@ -68,7 +75,7 @@
         </div>
     </fieldset>
 
-    @* Marketing Consent *@
+        @* Marketing Consent *@
     <fieldset class="fieldset">
 
         <div class="fieldset__heading">
@@ -95,28 +102,28 @@
     </fieldset>
 
 
-    @if(!IdentityGdprMarketingConsentSwitch.isSwitchedOn) {
-        @* Profiling *@
-        <fieldset class="fieldset">
-            <div class="fieldset__heading">
-                <h2 class="form__heading">Profiling</h2>
-            </div>
-            <div class="fieldset__fields">
-                <ul class="u-unstyled">
-                    <li class="form-field">
-                        <label class="label">In addition to the data that you provide to us, we may also match profiling data from third parties with your registration details.</label>
+        @if(!IdentityGdprMarketingConsentSwitch.isSwitchedOn) {
+            @* Profiling *@
+            <fieldset class="fieldset">
+                <div class="fieldset__heading">
+                    <h2 class="form__heading">Profiling</h2>
+                </div>
+                <div class="fieldset__fields">
+                    <ul class="u-unstyled">
+                        <li class="form-field">
+                            <label class="label">In addition to the data that you provide to us, we may also match profiling data from third parties with your registration details.</label>
 
-                        <div class="form-fields-group">
+                            <div class="form-fields-group">
                             @checkboxField(Checkbox(privacyForm("allowThirdPartyProfiling"), '_label -> "Allow matching with third party data"))(nonInputFields, messages)
-                        </div>
-                    </li>
-                </ul>
-            </div>
-        </fieldset>
-    }
+                            </div>
+                        </li>
+                    </ul>
+                </div>
+            </fieldset>
+        }
 
 
-    @* Submit button *@
+        @* Submit button *@
     <fieldset class="fieldset">
         <div class="fieldset__heading"></div>
         <div class="fieldset__fields">
@@ -128,6 +135,14 @@
         </div>
 
     </fieldset>
-</form>
+    </form>
+
+</div>
+
+
+<div>
+    @profile.emailPrefs(IdentityPage("/email-prefs", "Email preferences"), emailPrefsForm, emailSubscriptions, availableLists, idRequest, idUrlBuilder)
+</div>
+
 
 @registrationFooter(idRequest, idUrlBuilder)
